+++ src/Utils/API/CrystalMathLabs/Api.php
@@ -5,7 +5,7 @@ namespace App\Utils\API\CrystalMathLabs;
 use App\Utils\API\CrystalMathLabs\Endpoints\CurrentPlayerCount;
 use App\Utils\API\CrystalMathLabs\Endpoints\Stats;
 use App\Utils\API\CrystalMathLabs\Endpoints\TopPlayers;
-use App\Utils\API\CrystalMathLabs\Exceptions\ApiException;
+use App\Utils\API\ApiException;
 use App\Utils\Http\Session;
 use GuzzleHttp\Exception\RequestException;
