+++ src/Utils/API/CrystalMathLabs/Endpoints/CurrentPlayerCount.php
@@ -4,7 +4,7 @@ namespace App\Utils\API\CrystalMathLabs\Endpoints;
 
 use Psr\Http\Message\StreamInterface;
 
-class CurrentPlayerCount extends AbstractEndpointBase
+class CurrentPlayerCount extends AbstractEndpoint
 {
     protected string $end_point_url = 'type=players';
