+++ src/Utils/API/EndpointInterface.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Utils\API\CrystalMathLabs\Endpoints;
+namespace App\Utils\API;
 
-use App\Utils\API\CrystalMathLabs\Exceptions\ApiException;
 use GuzzleHttp\Exception\RequestException;
 use Psr\Http\Message\StreamInterface;
