--- src/Utils/API/OSRS/Exceptions/ApiException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Utils\API\OSRS\Exceptions;
-
-use Exception;
-
-class ApiException extends Exception {}
